Create Quicksort 2.0.c
In the code provided for quicksort, if the elements of input become same the program crashes because we weren't partitioning properly.
We were finding pivot according to the size of the array not taking a look at the array elements if there are elements smaller or greater than the pivot in left and right sides.
So, in the modified code, we have taken a function "partition" to make the element pivot which will be in the center of the array where left array elements will be smaller than the pivot and right elements will be greater. After that we will get a perfect array each time to run quicksort in the array.